✨ (DTOs): Change grade properties to double for decimal support ✨ (Database): Add Admin and SuperAdmin entities and configurations ✨ (Database): Seed initial Admin and SuperAdmin data ✨ (Entities): Add Admin and SuperAdmin entities ✨ (Entities): Update Professor entity with Gmail and OneTimeCode properties ✨ (Entities): Remove Average property from StudentStatistics entity ✨ (Entities): Change grade properties to double for decimal support ✨ (DI): Register AdminSeedData and SuperAdminSeedData The MidTermGrade and FinalGrade properties in GradeDto, SubjectDto, and SubjectStudentLink are changed from int to double to support decimal grades. New entities Admin and SuperAdmin are added with their configurations and seed data. Professor entity is updated with Gmail and OneTimeCode properties for authentication purposes. Average property is removed from StudentStatistics entity as it is not required. AdminSeedData and SuperAdminSeedData are registered in the dependency injection container.

+++ UniSphere/DTOs/Grades/GradeDto.cs
@@ -4,8 +4,8 @@ public sealed record GradeDto
 {
     public required int SubjectYear { get; init; }
     public required int SubjectSemester { get; init; }
-    public required int  MidTermGrade { get; init; }
-    public required int FinalGrade { get; init; }
+    public required double  MidTermGrade { get; init; }
+    public required double FinalGrade { get; init; }
     public required string SubjectName { get; init; }
     public required bool IsPassed { get; init; }
